backend/requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.30.1
sqlalchemy
asyncpg
psycopg2-binary==2.9.9
pg8000==1.30.3
alembic==1.13.2
celery==5.4.0
pydantic==2.8.2
pydantic-settings==2.4.0
PyJWT==2.8.0
pymilvus==2.4.4
pymongo==4.8.0
motor==3.5.0
PyMuPDF==1.24.7
pypandoc==1.13
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
python-docx==1.1.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pytz==2024.1
redis==5.0.7
requests==2.32.3

# Data Science & ML
ydata-profiling[default]
pandas
numpy
scikit-learn==1.5.1

# NLP and Text Processing
# For text summarization and keyword extraction, etc.
sumy==0.11.0
nltk==3.8.1
jieba==0.42.1
scikit-learn==1.5.1
transformers
langdetect==1.0.9
snownlp==0.12.3
sentence-transformers
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
ollama==0.3.0

# Image Processing
Pillow==10.4.0
imagehash==4.3.1

# Vector DB
pymilvus

# Graph DB
neo4j==5.22.0

# Monitoring
starlette-prometheus
python-json-logger

# Other
email_validator
fastapi-limiter
gunicorn
# uvloop
# httptools

# Testing
pytest
pytest-asyncio
httpx
requests
coverage
Faker
yarl==1.9.4
aiosqlite 